added test for personal_sendTransaction
