+++ tests/core/endpoints/test_personal_lock_account.py
@@ -1,51 +1,4 @@
-from eth_account import (
-    Account,
-)
-from eth_utils import (
-    encode_hex,
-)
 import pytest
-
-
-@pytest.fixture()
-def account_password():
-    return "a-password"
-
-
-@pytest.fixture()
-def account_private_key():
-    return Account().create().privateKey
-
-
-@pytest.fixture()
-def account_public_key(account_private_key):
-    return Account().privateKeyToAccount(account_private_key).address
-
-
-@pytest.fixture()
-def password_account(
-        rpc_client,
-        accounts,
-        account_password,
-        account_private_key,
-        account_public_key
-):
-    address = rpc_client(
-        'personal_importRawKey',
-        [encode_hex(account_private_key), account_password],
-    )
-    assert address == account_public_key
-
-    initial_balance = 1000000000000000000000  # 1,000 ether
-
-    rpc_client('eth_sendTransaction', [{
-        'from': accounts[0],
-        'to': address,
-        'value': initial_balance,
-    }])
-
-    assert rpc_client('eth_getBalance', [address]) == initial_balance
-    return address
 
 
 def test_personal_lockAccount(accounts, rpc_client, password_account, account_password):
